flask init
