+++ flaskapp/mac.owl
@@ -1,0 +1,94 @@
+<?xml version="1.0"?>
+<rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+         xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
+         xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
+         xmlns:owl="http://www.w3.org/2002/07/owl#"
+         xml:base="http://ifixit.org/mac.owl"
+         xmlns="http://ifixit.org/mac.owl#">
+
+<owl:Ontology rdf:about="http://ifixit.org/mac.owl"/>
+
+<owl:ObjectProperty rdf:about="#has_step">
+  <rdfs:domain rdf:resource="#Procedure"/>
+  <rdfs:range rdf:resource="#Step"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#in_toolbox">
+  <rdfs:domain rdf:resource="#Tool"/>
+  <rdfs:range rdf:resource="#Toolbox"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#has_image">
+  <rdfs:domain rdf:resource="#Step"/>
+  <rdfs:range rdf:resource="#Image"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#part_of">
+  <rdfs:domain rdf:resource="#Part"/>
+  <rdfs:domain rdf:resource="#Item"/>
+  <rdfs:range rdf:resource="#Item"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#used_in">
+  <rdfs:domain rdf:resource="#Tool"/>
+  <rdfs:range rdf:resource="#Step"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#has_procedure">
+  <rdfs:domain rdf:resource="#Part"/>
+  <rdfs:domain rdf:resource="#Item"/>
+  <rdfs:range rdf:resource="#Step"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#has_toolbox">
+  <rdfs:domain rdf:resource="#Procedure"/>
+  <rdfs:range rdf:resource="#Toolbox"/>
+</owl:ObjectProperty>
+
+<owl:ObjectProperty rdf:about="#subprocedure">
+  <rdfs:domain rdf:resource="#Procedure"/>
+  <rdfs:range rdf:resource="#Procedure"/>
+</owl:ObjectProperty>
+
+<owl:DatatypeProperty rdf:about="#step_number">
+  <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#FunctionalProperty"/>
+  <rdfs:domain rdf:resource="#Step"/>
+  <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#integer"/>
+</owl:DatatypeProperty>
+
+<owl:DatatypeProperty rdf:about="#step_description">
+  <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#FunctionalProperty"/>
+  <rdfs:domain rdf:resource="#Step"/>
+  <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#string"/>
+</owl:DatatypeProperty>
+
+<owl:Class rdf:about="#Procedure">
+  <rdfs:subClassOf rdf:resource="http://www.w3.org/2002/07/owl#Thing"/>
+</owl:Class>
+
+<owl:Class rdf:about="#Step">
+  <rdfs:subClassOf rdf:resource="#Procedure"/>
+</owl:Class>
+
+<owl:Class rdf:about="#Toolbox">
+  <rdfs:subClassOf rdf:resource="#Procedure"/>
+</owl:Class>
+
+<owl:Class rdf:about="#Tool">
+  <rdfs:subClassOf rdf:resource="#Toolbox"/>
+</owl:Class>
+
+<owl:Class rdf:about="#Item">
+  <rdfs:subClassOf rdf:resource="#Procedure"/>
+</owl:Class>
+
+<owl:Class rdf:about="#Part">
+  <rdfs:subClassOf rdf:resource="#Item"/>
+</owl:Class>
+
+<owl:Class rdf:about="#Image">
+  <rdfs:subClassOf rdf:resource="#Step"/>
+</owl:Class>
+
+
+</rdf:RDF>
