+++ flaskapp/templates/index.html
@@ -1,0 +1,15 @@
+
+{% extends "base.html" %}
+
+{% block title %}Browse{% endblock %}
+
+{% block header %}Welcome to iFixit{% endblock %}
+
+{% block content %}
+<div class="content">
+    <section class="home" id="home">
+        <h2>Home</h2>
+        <p>This is the home section where you can find an introduction or overview of the website.</p>
+    </section>
+</div>
+{% endblock %}
