+++ flaskapp/templates/search.html
@@ -1,0 +1,44 @@
+{% extends "base.html" %}
+
+{% block title %}Browse{% endblock %}
+
+{% block header %}Welcome to iFixit{% endblock %}
+
+{% block content %}
+<div class="content">
+    <section id="search">
+        <h2 class="search_header">Search Options</h2>
+        <div class="button-list">
+            {% for search in searches %}
+            <button onclick="performSearch('{{ search.id }}')"> {{ search.name }} </button>
+            {% endfor %}
+        </div>
+    </section>
+</div>
+
+<script>
+    async function performSearch(search_id) {
+        // Send a POST request with the search_id and search term
+        try {
+            console.log('Search ID:', search_id);
+            const response = await fetch('/get_query', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({ search_id: search_id })
+            });
+
+            if (!response.ok) {
+                throw new Error('Network response was not ok');
+            }
+
+            const result = await response.text();
+            // Print the returned string
+            alert('Search result: ' + result);
+        } catch (error) {
+            console.error('There was a problem with the fetch operation:', error);
+        }
+    }
+</script>
+{% endblock %}
